plugin: Added support for non sriov mode.
Two modes are supported.
1. sriov
This mode is useful when there is need to provision vhca per pod.
This vhca are based on sriov for IB and RoCE link layer.

2. hca
This mode is based on shared IB device among pods.
This is useful for RoCE macvlan mode.
This is seufl for IB mode when rdmacm is not used and users choice of
networking is used such as veth or vxlan etc.

Signed-off-by: Parav Pandit <parav@mellanox.com>